sphinx: Add support for transition indexing for older docs
Add support to create a 'transition' page for each Yocto Project
release. For each release we have a page that contains the links for
the released DocBook based documentation files.

How it works:
1. we create a new directive 'docbook-transition', and for
each old release we create an .rst file
(transition/<version>/index.rst), which uses this new include
directive and refer to a 'master' index file.
2. the new directive parses the 'master' index file, and replaces
@VER@ with the right version.
3. in conf.py, we add a hook, which is run at the end of the config,
and if 'srcdir' is transition/*/ which means we are building for a
legacy doc, we override version/release fields.
4. we create a new 'master' index file whenever we add or remove a
documentation file (e.g. whenever the template needs to be updated).

Also in Makefile we add a target ('publish') to generate a "built"
docs tree in a 'final' directory.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>