sphinx: ref-manual: use builtin glossary for the Terms section
Using the builting glossary is well suited for the Yocto Project
'terms' section. Conveniently, all terms will also be added in the
global index.

While converting this to a glossary, also fixed up some content which
was not properly converted by pandoc (such as codeblock sections, or
references in between terms).

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>