sphinx: ref-manual: fix heading for oe-init-build-env
In DocBook this section name is set to &OE_INIT_FILE;
variable. However using a substitution pattern in a heading in Sphinx
does not seem to work well. Let's just set the script environment name
here directly, without using a variable. We don't expect to change
this name anyways, so a variable is not really needed.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>