sphinx: fix up some trademark and branding issues
The following changes were required after a review of trademark and
branding guidelines:

1. add (R) to 'Yocto Project' on the top left (above the logo)
2. Fix up the capitalization of the  main page title
3. Add the copyright/legal blurb at the bottom of the page

For 3. it turned out to be simpler to override the whole footer.html
template, and maintain our own version. Also I took the liberty to
remove the 'next' and 'previous' buttons since I believe they are not
especially useful, given the navigation bar on the left side.

Reported-by: Tracey Erway <tracey.m.erway@intel.com>
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>