docs: ref-manual: ref-variables: add links to terms in glossary
Before the move to Sphinx, it used to be possible to get a direct link to a
variable from the term glossary. It is very useful when pointing people to
a specific variable when manually looking for it in the glossary.

Let's add this "feature" back.

Signed-off-by: Quentin Schulz <foss@0leil.net>
Reviewed-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
(cherry picked from commit 9e468274eaad270efd5f50e58a523798fcb8097e)
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>