Drop old presentations from repository
The `presentations` directory contained two talks which are now over 7
years old and are unlikely to be relevant to current users of the
project. They don't fit in well with the rest of the contents of this
repository and should be dropped. If anyone wants them, they'll still be
stored in the git history :)

Signed-off-by: Paul Barker <pbarker@konsulko.com>