documentation: Simplify yocto_wiki links
The `yocto_wiki` external link directive is modified to include the
`/wiki` element of the URL so that we can simplify the links in the
text.

Note that there are still a couple of places where this directive
cannot be used, such as in the table of contents in index.rst.

Signed-off-by: Paul Barker <pbarker@konsulko.com>
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>