ref-manual: reorder variable definitions
By alphabetical order, to get the same order as in the HTML
output, sorted thanks to the ":sorted:" directive.

That was needed after the recent variable renaming for
inclusive language, but other variables were already
out of order anyway.

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>
---

The change was made manually, but the source order was compared automatically
with the generated one thanks to these quick and dirty commands...

To extract the unsorted list of variable definitions from the sources:
> git grep "   :term:" documentation/ref-manual/variables.rst | grep -v "    " | cut -d ":" -f4  | cut -d "\`" -f2 > ~/tmp/unsorted

To extract the sorted list of variables from the generated HTML output:
> cd documentation
> make html
> egrep -o "^<dt id=\"term-([0-9A-Za-z\-_])*" _build/html/ref-manual/variables.html | cut -c 14- > ~/tmp/sorted

You'll most probably have better and simpler ideas for generating such lists!