set_versions/switchers: Drop versions shown to the active releases
I believe we should only be showing the "active" versions in the switchers
base list of releases to show. Zeus and warrior are old and no longer actively
maintained and we don't suggest new users use them. gatesgarth is also
outside it's support window.

I therefore propose removing these leaving us with dev, honister, hardknott,
dunfell (LTS). In addition, any release that is selected will be shown, along
with any later release in that release series.

People can still navigate to the older docs using the releases list or by
direct URL but this highlights to users which releases we'd expect/encourage
them to be using.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>