docs: set_versions.py: mark as obsolete only branches and old tags from obsolete releases
Branches are identified by their .999 version suffix which means they
will never be matched in the forloop above this git context. Therefore,
branches will match the condition. However, branches are not necessarily
obsolete (e.g. dunfell, honister and kirkstone today), so let's mark as
obsolete the branches which are from obsolete releases.

Old tags of currently supported releases are not defined as obsolete but
outdated, therefore using the series to which they belong like it is
done for branches is enough for obsolescence detection.

Cc: Quentin Schulz <foss+yocto@0leil.net>
Signed-off-by: Quentin Schulz <quentin.schulz@theobroma-systems.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>