overview-manual: licensing section fixes
Remove the reference to the GPL license, unnecessary
in this part of the manual and pointing to a wrong
link (LGPL instead of GPL).

Anyway, we have no authority to say that the MIT license
is compatible with the GPL.

Also provide details about the MIT license through Wikipedia,
instead of just the raw license text with no explanations
through opensource.org.

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>