docs: conf.py: bump minimum Sphinx version requirement
extlinks captions support using %s substitution but only from sphinx
4.0 onwards. Anything before does not support the %s substitution.
However, this will be a requirement in sphinx 6.0 and later, and sphinx
5.0 already shows a warning telling us that.

Since the build is made with -W flag which fails it if any warning is
printed, a fix is required.

Sphinx 4.0.0 was released two years ago. Unfortunately, Debian packages
are out-of-date and only the latest Ubuntu has recent enough packages.
Therefore, pip installation should be recommended now instead and the
minimum version for building the docs bumped to 4.0.

Cc: Quentin Schulz <foss+yocto@0leil.net>
Signed-off-by: Quentin Schulz <quentin.schulz@theobroma-systems.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>