docs: auto-generate releases.rst
In order to maintain one less file, let's auto-generate the releases.rst
file which contains a link for each release ever released.

This gets auto-generated by checking the tags available in this git repo
and adding a link for each that exists.

A few tags are notoriously missing from this git repo and they are
manually listed then, until the tags are pushed for the appropriate
commit.

Cc: Quentin Schulz <foss@0leil.net>
Signed-off-by: Quentin Schulz <quentin.schulz@theobroma-systems.com>
Reviewed-by: Michael Opdenacker <michael.opdenacker@bootlin.com>

---

Changes in V2:
 * set_versions.py: update description
 * set_versions.py: add Quentin Schulz to the list of authors
 * set_versions.py: remove trailing whitespace in generated
   section header (harmless but flagged by vim)