ref-manual: faq.rst: reorganize into subsections, contents at top
Reorganize the FAQ into sections and subsections, which allows
to have a table of contents at the top. This makes it easier
to find relevant questions without having to scroll down the entire page.

Reduce the size of questions so that they fit in a subsection title.

Merged two redundant questions about fetching sources and working
through proxies.

Minor updates to other sections.

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>