ref-manual: classes.rst: add links to all references to a class
[YOCTO #14508]
Reported-by: Quentin Schulz <foss@0leil.net>
Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>

--

Doing this to consistently replace any reference to a class
by the corresponding link. This is a bit trivial within the declaration
of a class, but helps making sure that this rule applies everywhere.

This helps for example to rename or remove classes from the
documentation.

See https://bugzilla.yoctoproject.org/show_bug.cgi?id=14508

As this change is time consuming, submitting the first part of it
as an RFC, to double check there is an agreement on doing this.