docs: poky.yaml.in: remove pylint3 from Ubuntu/Debian host dependencies
It is not actually needed outside of the autobuilder and since the
package got renamed in Ubuntu 20.04 and later to pylint, it was giving
the users an incorrect command line to run and would fail all the time,
better remove it instead of trying to handle Debian/recent Ubuntu
differences in the documentation.

Moreover, the other supported host distributions didn't have pylint
listed as a dependency, proving that it's indeed not needed.

Cc: Quentin Schulz <foss+yocto@0leil.net>
Signed-off-by: Quentin Schulz <quentin.schulz@theobroma-systems.com>
Reviewed-by: Michael Opdenacker <michael.opdenacker@bootlin.com>