ref-manual: add references to variables only documented in the BitBake manual
This makes the Yocto Project manual variable list and index more useful,
not missing any potentially useful variable.

The price to pay is that a reference to such variables now has to go through
the intermediate reference in the Yocto Project variable list, instead
of going straight to the Bitbake manual.

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>
Reviewed-by: Quentin Schulz <foss+yocto@0leil.net>