manuals: fix double colons
Fixing double colons appearing alone on a line, while
they could be put at the end of the previous line.

Sometimes placing a note after the quoted text
to avoid such a situation. It's more natural too
not to have a note between the introduction text
and the quoted section.

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>

---

Note: instances to replace detected by the following command:
git grep "^\s*\:\:\s*$"