manuals: standards.md5: add standard for project names
Set a new standard to introduce project names with single quotes,
so that they appear in generated text with an italic font,
to make them easier to distinguish from command names and from
ordinary English words.

Rework and move the standard for command and file names
to make the whole description "flow" better.

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>
CC: Quentin Schulz <quentin.schulz@theobroma-systems.com>

---

Changes in V4
- Remove the exception for references to section titles (rare case)
- Remove the exception for single quotes in section titles
  (not needed are single quotes are properly processed in this case)

Changes in V3
- Mention that project names don't need single quotes
  when they are part of a hyperlink.

Changes in V2
- Specify that uncapitalized project names should remain so
  even at the beginning of a sentence. This follows what Wikipedia
  seems to be doing, e.g. on https://en.wikipedia.org/wiki/Perf_(Linux)